move fixture to conftest
