+++ source/databricks/tests/geh_stream/helpers/test_schemas/charges_with_price_and_links_schema.py
